+++ lib/std/strings.c
@@ -1,16 +1,16 @@
 #include <data/types.h>
 
-unsigned int str_len(const char* str)
+size_t str_len(const char* str)
 {
-	unsigned int len = 0;
+	size_t len = 0;
 	while (*(str++))
 		len++;
 	return len;
 }
 
-char* strncpy(char* dest, const char* src, uint64 n)
+char* strncpy(char* dest, const char* src, uint32 n)
 {
-	uint64 i;
+	uint32 i;
 	for (i = 0; i < n && src[i] != '\0'; ++i)
 		dest[i] = src[i];
 	for (; i < n; ++i)
